♻️ use MessageList comp. in ChatScreen
